Move aggregate cracking time from TaskUtils to AgentUtils, refactor other code accordingly
