+++ ci/phpunit/inc/utils/TaskUtilsTest.php
@@ -250,65 +250,4 @@ final class TaskUtilsTest extends TestBase {
     $taskUpdated = Factory::getTaskFactory()->get($taskObjects["task"]->getId());
     $this->assertEquals(0, $taskUpdated->getIsCpuTask());
   }
-  
-  
-  /**
-   * Test cracking time aggregation for an agent on a task.
-   *
-   * @return void
-   * @throws Exception
-   */
-  public function testCrackingTimeAggregation(): void {
-    $task = $this->createTaskHelper()["task"];
-    $agent1 = $this->createAgent("test");
-    $agent2 = $this->createAgent("test");
-    $timeSpans = (array) [
-      [1000, 2000, $agent1],
-      [3000, 5000, $agent2],
-      [3000, 8000, $agent1],
-      [8000, 10000, $agent1],
-      [15000, 20000, $agent1],
-    ];
-    
-    foreach ($timeSpans as [$start, $end, $agent]) {
-      $chunk = $this->createChunk($task, $agent, 4);
-      $chunk->setDispatchTime($start);
-      $chunk->setSolveTime($end);
-      Factory::getChunkFactory()->update($chunk);
-    }
-    
-    // Calculate reference value via an interval merge algorithm
-    $totalEnd = $referenceSum = 0;
-    usort($timeSpans, fn($a, $b) => $a[0] <=> $b[0]); // Make sure time spans are sorted
-    foreach ($timeSpans as [$currentStart, $currentEnd, $agent]) { // Expects list to be sorted by time
-      if ($agent == $agent1) {
-        $referenceSum = $referenceSum + ($currentEnd - $currentStart) // Add current time span to running total
-          - max(0, $totalEnd - $currentStart) // Correct for potential overlapping when current start before previous end
-          + max(0, $totalEnd - $currentEnd); // Correct for potential overcorrection when current end before previous end
-        $totalEnd = max($totalEnd, $currentEnd); // Extend window if current end exceeds previous end
-      }
-    }
-    
-    // Calculate aggregate cracking time via TaskUtils
-    $crackingTime = TaskUtils::getAggregateCrackingTime($agent1->getId(), $task->getId());
-    
-    $this->assertEquals($referenceSum, $crackingTime);
-  }
-
-  public function createTaskHelper(): array {
-    $user = $this->createUser("phpunit");
-    $accessGroup = $this->createAccessGroup("phpunit");
-    $this->createAccessGroupUser($user, $accessGroup);
-
-    $hashType = $this->createHashType();
-    $hashlist = $this->createHashlist($accessGroup, $hashType);
-
-    $taskWrapper = $this->createTaskWrapper($accessGroup, $hashlist);
-
-    $crackerBinaryType = $this->createCrackerBinaryType();
-    $crackerBinary = $this->createCrackerBinary($crackerBinaryType);
-    $task = $this->createTask($taskWrapper, $crackerBinary, $crackerBinaryType);
-
-    return array("user"=> $user, "accessGroup"=>$accessGroup, "hashType"=>$hashType, "hashlist"=>$hashlist, "taskWrapper"=>$taskWrapper, "crackerBinaryType"=>$crackerBinaryType, "crackerBinary"=>$crackerBinary, "task"=>$task);
-  }
 }
